update: extract utils (wip: zodify utils)
